Revert "Have the uap version of RemoteExecutorTestBase.RemoteInvoke wait in a… (#21780)"
This reverts commit 066ad32f1352fe5325998b997b93a0e0cb54cf58.
