+++ src/Common/tests/System/Threading/ThreadTestHelpers.cs
@@ -4,9 +4,8 @@
 
 using Xunit;
 
-namespace System.Threading
+namespace System.Threading.Tests
 {
-    // NOTE: This project has a ref assembly, any changes to public members of this class will require changes to the ref as well
     public static class ThreadTestHelpers
     {
         public const int ExpectedTimeoutMilliseconds = 50;
