+++ src/System.Threading.ThreadPool/tests/ThreadPoolTests.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Threading.Tasks;
+using System.Threading.Tests;
 using Xunit;
 
 namespace System.Threading.ThreadPools.Tests
